requirements.txt
# Core dependencies for LUC-CMFD

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.19.0
scikit-learn>=1.0.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=9.0.0

# Data Processing
pandas>=1.3.0
pyyaml>=6.0

# Utilities
tqdm>=4.62.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Jupyter (for notebook)
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0
